refactor: inline Protocol data fetching, remove wrapper
Eliminates the ProtocolCards wrapper that existed only to fetch data
and show skeletons. The Protocol page now uses the shared ProtocolCards
component directly, avoiding a naming collision and an unnecessary
indirection layer.
